pyproject.toml
[project]
name = "run-watcher"
version = "0.1.0"
description = "A Textual TUI app for monitoring GitHub Actions runs"
requires-python = ">=3.10"
dependencies = [
    "textual>=0.47.0",
    "PyGithub>=2.1.1",
    "python-dotenv>=1.0.0",
]

[project.scripts]
run-watcher = "run_watcher.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["run_watcher"]
